refactor: ResultCardコンポーネントを抽出しカードヘッダーを統一
TableContentとChartCardで重複していたカード枠+ヘッダー描画を
共通のResultCardに集約。ChartCardでweight未考慮だったn値表示も修正。
