+++ src/components/aggregation/ResultView.tsx
@@ -1,13 +1,12 @@
 import { useEffect, useState } from "preact/hooks";
 import type { AggResult } from "../../lib/agg/aggregate";
 import { pivot } from "../../lib/agg/pivot";
-import { resolveQuestionLabel } from "../../lib/labels";
 import { Toolbar, ViewOpts, type PctDirection, type ViewMode } from "./Toolbar";
 import { GtTable } from "./GtTable";
 import { CrossTable } from "./CrossTable";
 import { AIBubble } from "./AIBubble";
 import { ChartCard, type GtChartType } from "./ChartRenderer";
-import { useAggregation } from "./AggregationContext";
+import { ResultCard } from "./ResultCard";
 
 export interface ResultViewProps {
   results: AggResult[];
@@ -120,7 +119,6 @@ function TableContent({
   hasCross: boolean;
   pctDirection: PctDirection;
 }) {
-  const { layoutMeta, weightCol } = useAggregation();
   const allGtCells = results.flatMap((r) => r.cells.filter((c) => c.sub === "GT"));
   const maxPct = Math.max(...allGtCells.map((c) => c.pct), 0);
 
@@ -136,33 +134,18 @@ function TableContent({
         const pv = pivot(res.cells);
         const isCross = pv.subs.length > 1;
 
-        const gtSub = pv.subs.find((s) => s.label === "GT")!;
-        const nLabel = weightCol ? `n=${gtSub.n.toFixed(1)}` : `n=${gtSub.n.toLocaleString()}`;
-
-        const questionLabel = resolveQuestionLabel(res.question, layoutMeta);
-        const hasLabel = questionLabel !== res.question;
-
         return (
-          <div
-            class={`overflow-hidden rounded-xl border border-border bg-surface shadow-sm${isCross ? " overflow-x-auto" : ""}`}
+          <ResultCard
             key={res.question}
+            res={res}
+            extraClass={isCross ? "overflow-x-auto" : undefined}
           >
-            <div class="flex items-baseline gap-3 border-b border-border p-4">
-              <div class="flex min-w-0 flex-col gap-[2px]">
-                <span class="text-[0.875rem] font-bold text-accent">{questionLabel}</span>
-                {hasLabel && (
-                  <span class="text-xs tracking-[0.04em] text-muted">{res.question}</span>
-                )}
-              </div>
-              <span class="text-xs tracking-[0.04em] text-muted">{res.type}</span>
-              <span class="ml-auto text-[0.8125rem] text-muted">{nLabel}</span>
-            </div>
             {isCross ? (
               <CrossTable res={res} pv={pv} pctDir={pctDirection} />
             ) : (
               <GtTable res={res} pv={pv} maxPct={maxPct} />
             )}
-          </div>
+          </ResultCard>
         );
       })}
     </div>
